requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.1
discord.py==1.7.3
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
python-dotenv==0.19.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
yarl==1.6.3
